perf: reduce export memory usage with 16-bit PCM and OPFS streaming output (#506)
* feat: store decoded audio as 16-bit PCM


* fix: propagate encoder and muxer errors to the backpressure wait


* feat: stream export output to OPFS instead of buffering in memory


* refactor: extract shared per-frame FFT computation


* test: add memory benchmark comparing in-memory and OPFS export targets


* chore: remove redundant comments


* fix: address Copilot review feedback

Propagate the abort reason to the OPFS writable and stop any existing
heap poller before starting a new one.


---------

+++ src/lib/audio/audio.ts
@@ -3,7 +3,7 @@ export interface SerializedAudio {
   readonly sampleRate: number;
   readonly numberOfChannels: number;
   readonly duration: number;
-  readonly channels: Float32Array[];
+  readonly channels: Int16Array[];
 }
 
 /** Serializable audio data for IndexedDB storage (duration is derived) */
